% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_horizons.R
\name{error_horizons}
\alias{error_horizons}
\title{get the accuracy of certain section}
\usage{
error_horizons(error, start, end)
}
\arguments{
\item{error}{A list of the error.}

\item{start}{start of the section you want to choose (start with 1).}

\item{end}{end of the section you want to choose (start with 1).}
}
\value{
average of the certain section
}
}
\description{
get the accuracy of certain section
}
