% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_series.r
\name{forecast_series}
\alias{forecast_series}
\title{forecast the coming time series}
\usage{
forecast_series(
  dataset,
  method = c("combine", "aic", "fb"),
  features = NULL,
  error_model = NULL,
  trend_model = NULL,
  seasonality_model = NULL
)
}
\arguments{
\item{dataset}{A list the elements having a \code{ts} object with the name \code{x}.}

\item{method}{the way you want to forecast the series(aic,feature based or combine the two methods).}

\item{features}{NULL if you just use the model package provides.}

\item{error_model}{the error model.}

\item{trend_model}{the trend model.}

\item{seasonality_model}{the seasonality model.}
}
\value{
A list with the elements having the following structure
}
\description{
forecast the coming time series
}
