% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.r
\name{train_model}
\alias{train_model}
\title{train model with new features.}
\usage{
train_model(train_features)
}
\arguments{
\item{train_features}{the features for the generated time series.}
}
\value{
three classifer of each ets component
}
\description{
For each series in \code{dataset}, point and interval forecasting performance
are evaluated for all methods in \code{methods} of \code{\link{ts_forec}}
with regard to MASE, sMAPE, MSIS, Spread, Coverage and Upper coverage.
}
